........
  r1153049 | davsclaus | 2011-08-02 03:14:03 -0400 (Tue, 02 Aug 2011) | 1 line
  
  Fixed problem with attribute not using same type in ManagedThreadPool causing that option not accessible in JMX
........


git-svn-id: https://svn.apache.org/repos/asf/camel/branches/camel-2.8.x@1172775 13f79535-47bb-0310-9956-ffa450edef68

+++ camel-core/src/main/java/org/apache/camel/management/mbean/ManagedThreadPool.java
@@ -149,7 +149,7 @@ public class ManagedThreadPool {
     }
 
     @ManagedAttribute(description = "Keep alive time in seconds")
-    public void setKeepAliveTime(int keepAliveTimeInSeconds) {
+    public void setKeepAliveTime(long keepAliveTimeInSeconds) {
         threadPool.setKeepAliveTime(keepAliveTimeInSeconds, TimeUnit.SECONDS);
     }
